+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.3.0 - 9/15/2020
+* Removed boost dependency
+* Added kinematic hardening to crystal plasticity model
+* Added work-based damage model
+* Block conversion routines for going to/from tensor notation
+* Updated build/release process
+
 ## 1.2.2 - 4/7/2020
 * Bugfix for new python package distribution